Optimization commented out
Optimization is not working. I am very close.

Lines 168 - 177 of graphics.s and the checkPicture in the graphics.s are going to be where to look at it.


The game is relatively stable currently. The finish state needs to be added. As well - when the car is hit another car, reset the car